# src/videos/dto/video.input-dto.ts
import { Resolutions } from "../types/videos";


export type UpdateVideoInputModel = {

   title: string;
   author: string;
   availableResolutions: Resolutions[];
   canBeDownloaded: boolean; 
   minAgeRestriction: number | null;
   publicationDate: Date;

}

export type CreateVideoInputModel = {

   title: string;
   author: string;
   availableResolutions: Resolutions[];
}